requirements.txt
requests
bs4
pytz
pixivpy3
python-dotenv
selenium
numpy
tensorflow
natsort
pillow
matplotlib
opencv-contrib-python
scikit-learn